refactor(sputnik): flatten llrt package for utils
