--- src/sputnik/src/js/apis/node/blob/llrt/mod.rs
@@ -1,6 +1,0 @@
-mod file;
-pub mod polyfill;
-#[rustfmt::skip]
-mod utils;
-
-pub use utils::primordials::{BasePrimordials, Primordial};
