+++ src/sputnik/src/js/apis/node/llrt/buffer/file.rs
@@ -1,12 +1,12 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-use crate::js::apis::node::blob::llrt::utils::time;
+use crate::js::apis::node::llrt::utils::time;
 use rquickjs::{
     atom::PredefinedAtom, class::Trace, function::Opt, ArrayBuffer, Coerced, Ctx, Exception,
     IntoJs, Object, Result, Value,
 };
 
-use super::polyfill::Blob;
+use super::blob::Blob;
 
 #[rquickjs::class]
 #[derive(Trace, Clone, rquickjs::JsLifetime)]
