+++ src/sputnik/src/js/apis/node/llrt/utils/bytes.rs
