+++ src/sputnik/src/js/apis/node/llrt/utils/class.rs
