+++ src/sputnik/src/js/apis/node/llrt/utils/object.rs
