+++ src/sputnik/src/js/apis/node/llrt/utils/result.rs
