+++ src/sputnik/src/js/apis/node/mod.rs
@@ -1,7 +1,7 @@
-mod blob;
+mod llrt;
 mod javy;
 
-use crate::js::apis::node::blob::init_blob;
+use crate::js::apis::node::llrt::init_blob;
 use crate::js::apis::node::javy::init_text_encoding;
 use rquickjs::{Ctx, Error as JsError};
